aiService/requirements.txt
# FastAPI sunucusu için
fastapi
uvicorn[standard]
python-multipart
celery
redis
google-generativeai
PyPDF2
zemberek-python
zeyrek
httpx
pydantic-settings  
python-dotenv
gTTS
ollama
nltk
requests
pymupdf
pillow